requirements-lambda.txt
attrs==21.4.0
aws-cdk-lib==2.12.0
black==22.1.0
cattrs==1.10.0
click==8.0.4
constructs==10.0.82
importlib-metadata==4.11.2
iniconfig==1.1.1
jsii==1.54.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
publication==0.0.3
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.2
typing_extensions==4.1.1
zipp==3.7.0
